+++ src/token.rs
@@ -1,0 +1,156 @@
+#[derive(Debug, Clone)]
+pub struct Token {
+    pub kind: Kind,
+    // Value might be deleted in the future
+    pub value: TokenValue,
+    pub start: usize,
+    pub end: usize,
+}
+
+// TODO: remove this after implementing all the tokens
+#[allow(dead_code)]
+// https://docs.python.org/3/reference/lexical_analysis.html
+#[derive(Debug, PartialEq, Clone, Copy)]
+pub enum Kind {
+    // Line structure
+    NewLine, // \n
+    Indent,  // \t
+    Dedent,  // \t
+
+    // Identifiers
+    Identifier,
+
+    // Keywords
+    False,    // False
+    None,     // None
+    True,     // True
+    And,      // and
+    As,       // as
+    Assert,   // assert
+    Async,    // async
+    Await,    // await
+    Break,    // break
+    Class,    // class
+    Continue, // continue
+    Def,      // def
+    Del,      // del
+    Elif,     // elif
+    Else,     // else
+    Except,   // except
+    Finally,  // finally
+    For,      // for
+    From,     // from
+    Global,   // global
+    If,       // if
+    Import,   // import
+    In,       // in
+    Is,       // is
+    Lambda,   // lambda
+    Nonlocal, // nonlocal
+    Not,      // not
+    Or,       // or
+    Pass,     // pass
+    Raise,    // raise
+    Return,   // return
+    Try,      // try
+    While,    // while
+    With,     // with
+    Yield,    // yield
+
+    // String Literals
+    StringLiteral,
+    // f-string literal
+    FString, // f or F prefix
+    // byte string literal
+    Bytes, // b or B prefix
+    // raw string literal
+    RawString, // r or R prefix
+    // raw f-string literal,
+    RawFString, // rf or rF or Rf or RF or fr or fR or Fr or FR prefix
+    // raw byte string literal
+    RawBytes, // rb or rB or Rb or RB or br or bR or Br or BR prefix
+    // unicode string literal,
+    Unicode, // u or U prefix
+
+    Integer,                // 123
+    Binary,                 // 0b1010
+    Octal,                  // 0o123
+    Hexadecimal,            // 0x123
+    PointFloat,             // 123.456
+    ExponentFloat,          // 123.456e-10
+    ImaginaryInteger,       // 123j
+    ImaginaryPointFloat,    // 123.2j
+    ImaginaryExponentFloat, // 123.456e-10j
+
+    // Operators
+    Plus,       // +
+    Minus,      // -
+    Mul,        // *
+    Pow,        // **
+    Div,        // /
+    IntDiv,     // //
+    Mod,        // %
+    MatrixMul,  // @
+    LeftShift,  // <<
+    RightShift, // >>
+    BitAnd,     // &
+    BitOr,      // |
+    BitXor,     // ^
+    BitNot,     // ~
+    Walrus,     // :=
+    Less,       // <
+    Greater,    // >
+    LessEq,     // <=
+    GreaterEq,  // >=
+    Eq,         // ==
+    NotEq,      // !=
+
+    // Delimiters
+    LeftParen,        // (
+    RightParen,       // )
+    LeftBrace,        // [
+    RightBrace,       // ]
+    LeftBracket,      // {
+    RightBracket,     // }
+    Comma,            // ,
+    Colon,            // :
+    Dot,              // .
+    SemiColon,        // ;
+    Assign,           // =
+    Arrow,            // ->
+    AddAssign,        // +=
+    SubAssign,        // -=
+    MulAssign,        // *=
+    DivAssign,        // /=
+    ModAssign,        // %=
+    MatrixMulAssign,  // @=
+    BitAndAssign,     // &=
+    BitOrAssign,      // |=
+    BitXorAssign,     // ^=
+    IntDivAssign,     // //=
+    ShiftLeftAssign,  // <<=
+    ShiftRightAssign, // >>=
+    PowAssign,        // **=
+
+    // Special
+    Sharp,     // #
+    BackSlash, // \
+
+    // Unconditional error
+    Dollar,       // $
+    QuestionMark, // ?
+    BackTick,     // `
+
+    // Others
+    Comment,
+    WhiteSpace,
+    Eof,
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub enum TokenValue {
+    None,
+    Number(String), // TODO: String because we don't need the value yet
+    Str(String),
+    Indent(usize),
+}
